Update reviewer list when review is submitted from the description page
