+++ preview-src/pullRequestOverviewRenderer.ts
@@ -5,7 +5,7 @@
 
 import { dateFromNow } from '../src/common/utils';
 import { TimelineEvent, CommitEvent, ReviewEvent, CommentEvent, isCommentEvent, isReviewEvent, isCommitEvent, isMergedEvent, MergedEvent } from '../src/common/timelineEvent';
-import { PullRequestStateEnum, MergeMethod, MergeMethodsAvailability } from '../src/github/interface';
+import { PullRequestStateEnum, ReviewState, MergeMethodsAvailability, ILabel, MergeMethod } from '../src/github/interface';
 import md from './mdRenderer';
 import { MessageHandler } from './message';
 import { getState, updateState, PullRequest } from './cache';
@@ -14,9 +14,14 @@ import { Comment } from '../src/common/comment';
 const commitIconSvg = require('../resources/icons/commit_icon.svg');
 const mergeIconSvg = require('../resources/icons/merge_icon.svg');
 const editIcon = require('../resources/icons/edit.svg');
-const deleteIcon = require('../resources/icons/delete.svg');
 const checkIcon = require('../resources/icons/check.svg');
-const dotIcon = require('../resources/icons/dot.svg');
+
+const plusIcon = require('../resources/icons/plus.svg');
+const deleteIcon = require('../resources/icons/delete.svg');
+
+const pendingIcon = require('../resources/icons/dot.svg');
+const commentIcon = require('../resources/icons/comment.svg');
+const diffIcon = require('../resources/icons/diff.svg');
 
 const emoji = require('node-emoji');
 
@@ -93,7 +98,7 @@ function getStateIcon(state: string) {
 	} else if (state === 'failure') {
 		return deleteIcon;
 	} else {
-		return dotIcon;
+		return pendingIcon;
 	}
 }
 
@@ -1001,7 +1006,7 @@ class ReviewNode {
 				command: buttonAction,
 				args: commentingArea.value
 			}).then(message => {
-				appendReview(message.value, this._messageHandler);
+				appendReview(message, this._messageHandler);
 			}, err => {
 				// Handle error
 				submitButton.disabled = false;
@@ -1021,7 +1026,169 @@ class ReviewNode {
 	}
 }
 
-export function appendReview(review: ReviewEvent, messageHandler: MessageHandler): void {
+function renderSection(containerId: string, label: string, addCommand: string, renderItems: (newItems?: any[]) => HTMLElement[], messageHandler): void {
+	const container = document.getElementById(containerId);
+	container.innerHTML = '';
+
+	const sectionLabel = document.createElement('div');
+	sectionLabel.className = 'section-header';
+
+	const sectionText = document.createElement('div');
+	sectionText.textContent = label;
+	sectionLabel.appendChild(sectionText);
+
+	const addButton = document.createElement('button');
+	addButton.innerHTML = plusIcon;
+	addButton.title = `Add ${label}`;
+	addButton.addEventListener('click', () => {
+		messageHandler.postMessage({
+			command: addCommand
+		}).then(message => {
+			const updatedItems = renderItems(message.added);
+			sectionContent.innerHTML = '';
+			updatedItems.forEach(item => sectionContent.appendChild(item));
+		});
+	});
+	sectionLabel.appendChild(addButton);
+
+	const sectionContent = document.createElement('div');
+	sectionContent.className = 'section-content';
+
+	container.appendChild(sectionLabel);
+	container.appendChild(sectionContent);
+
+	const items = renderItems();
+	items.forEach(item => sectionContent.appendChild(item));
+}
+
+function getReviewStateElement(state: string): HTMLElement {
+	const reviewState = document.createElement('div');
+
+	switch (state) {
+		case 'REQUESTED':
+			reviewState.innerHTML = pendingIcon;
+			reviewState.title = 'Awaiting requested review';
+			break;
+		case 'COMMENTED':
+			reviewState.innerHTML = commentIcon;
+			reviewState.title = 'Left review comments';
+			break;
+		case 'APPROVED':
+			reviewState.innerHTML = checkIcon;
+			reviewState.title = 'Approved these changes';
+			break;
+		case 'CHANGES_REQUESTED':
+			reviewState.innerHTML = diffIcon;
+			reviewState.title = 'Requested changes';
+			break;
+	}
+	return reviewState;
+}
+
+export function renderDeleteButton(label: string, command: string, args: any, update: () => void, messageHandler: MessageHandler): HTMLElement {
+	const deleteButton = document.createElement('button');
+	deleteButton.innerHTML = deleteIcon;
+	deleteButton.children[0].classList.add('hidden-focusable');
+	deleteButton.title = `Remove ${label}`;
+	deleteButton.addEventListener('click', () => {
+		messageHandler.postMessage({ command, args }).then(_ => update());
+	});
+
+	deleteButton.addEventListener('focus', () => {
+		deleteButton.children[0].classList.remove('hidden-focusable');
+	});
+
+	deleteButton.addEventListener('blur', () => {
+		deleteButton.children[0].classList.add('hidden-focusable');
+	});
+
+	return deleteButton;
+}
+
+export function renderReviewers(pr: PullRequest, messageHandler: MessageHandler): void {
+	renderSection('reviewers', 'Reviewers',
+		'pr.add-reviewers',
+		(newItems?: ReviewState[]): HTMLElement[] => {
+			if (newItems) {
+				pr.reviewers = pr.reviewers.concat(newItems);
+				updateState({ reviewers: pr.reviewers });
+			}
+
+			return pr.reviewers.map((reviewer, i) => {
+				const reviewerElement = document.createElement('div');
+				reviewerElement.classList.add('section-item', 'reviewer');
+
+				const userIcon = renderUserIcon(reviewer.reviewer.url, reviewer.reviewer.avatarUrl);
+				reviewerElement.appendChild(userIcon);
+
+				const userName = document.createElement('div');
+				userName.className = 'login';
+				reviewerElement.appendChild(userName);
+				userName.textContent = reviewer.reviewer.login;
+
+				const reviewState = getReviewStateElement(reviewer.state);
+				reviewerElement.appendChild(reviewState);
+
+				if (reviewer.state === 'REQUESTED') {
+					const deleteButton = renderDeleteButton('reviewer', 'pr.remove-reviewer', reviewer.reviewer.login, () => {
+						pr.reviewers.splice(i, 1);
+						updateState({ reviewers: pr.reviewers });
+						reviewerElement.remove();
+					}, messageHandler);
+					reviewerElement.appendChild(deleteButton);
+
+					reviewerElement.addEventListener('mouseover', () => {
+						deleteButton.children[0].classList.remove('hidden-focusable');
+					});
+
+					reviewerElement.addEventListener('mouseout', () => {
+						if (document.activeElement !== deleteButton) {
+							deleteButton.children[0].classList.add('hidden-focusable');
+						}
+					});
+				}
+
+				return reviewerElement;
+			});
+	}, messageHandler);
+}
+
+export function renderLabels(pr: PullRequest, messageHandler: MessageHandler): void {
+	renderSection('labels', 'Labels',
+		'pr.add-labels',
+		(newItems?: ILabel[]): HTMLElement[] => {
+			if (newItems) {
+				pr.labels = pr.labels.concat(newItems);
+				updateState({ labels: pr.labels });
+			}
+
+			return pr.labels.map((label, i) => {
+				const labelElement = document.createElement('div');
+				labelElement.textContent = label.name;
+				labelElement.classList.add('label', 'section-item');
+
+				const deleteButton = renderDeleteButton('label', 'pr.remove-label', label.name, () => {
+					pr.labels.splice(i, 1);
+					updateState({ labels: pr.labels });
+					labelElement.remove();
+				}, messageHandler);
+				labelElement.appendChild(deleteButton);
+				labelElement.addEventListener('mouseover', () => {
+					deleteButton.children[0].classList.remove('hidden-focusable');
+				});
+
+				labelElement.addEventListener('mouseout', () => {
+					if (document.activeElement !== deleteButton) {
+						deleteButton.children[0].classList.add('hidden-focusable');
+					}
+				});
+
+				return labelElement;
+			});
+		}, messageHandler);
+}
+
+export function appendReview(data: { review: ReviewEvent, reviewers: ReviewState[] }, messageHandler: MessageHandler): void {
 	const state = getState();
 
 	let events = state.events;
@@ -1034,10 +1201,12 @@ export function appendReview(review: ReviewEvent, messageHandler: MessageHandler
 		});
 	}
 
-	events.push(review);
-	updateState({ events: events });
+	events.push(data.review);
+	state.reviewers = data.reviewers;
+	updateState({ events: events, reviewers: state.reviewers });
 
 	renderTimelineEvents(state, messageHandler);
+	renderReviewers(state, messageHandler);
 
 	clearTextArea();
 }
